improve keybinds on mac
